+++ src/lib/components/Form/validators/IsEqualValidator.ts
@@ -1,0 +1,1 @@
+export const IsEqualValidator = (valueA: any, valueB: any) => valueA === valueB;