+++ src/lib/components/Form/validators/index.ts
@@ -1,2 +1,3 @@
 export * from './EmailValidtor';
 export * from './IsEmptyValidator';
+export * from './IsEqualValidator';